+++ packages/editor/editor-core/src/plugins/extension/ui/Extension/InlineExtension/index.tsx
@@ -10,16 +10,23 @@ export interface Props {
   node: PmNode;
   macroProvider?: MacroProvider;
   onClick: (event: React.SyntheticEvent<any>) => void;
+  children?: React.ReactNode;
 }
 
 export default class InlineExtension extends Component<Props, any> {
   render() {
-    const { node, onClick } = this.props;
+    const { node, onClick, children } = this.props;
+
+    const hasChildren = !!children;
+
+    const className = hasChildren
+      ? 'with-overlay with-children'
+      : 'with-overlay';
 
     return (
-      <Wrapper onClick={onClick} className="with-overlay">
+      <Wrapper onClick={onClick} className={className}>
         <Overlay className="extension-overlay" />
-        <ExtensionLozenge node={node} />
+        {children ? children : <ExtensionLozenge node={node} />}
       </Wrapper>
     );
   }