+++ packages/editor/editor-test-helpers/src/schema-builder.ts
@@ -4,6 +4,7 @@ import {
   MediaSingleAttributes,
   ApplicationCardAttributes,
   CellAttributes,
+  LinkAttributes,
 } from '@atlaskit/editor-common';
 import {
   Fragment,
@@ -370,7 +371,7 @@ export const code = markFactory(sampleSchema.marks.code, {});
 export const strike = markFactory(sampleSchema.marks.strike, {});
 export const mentionQuery = (attrs = { active: true }) =>
   markFactory(sampleSchema.marks.mentionQuery, attrs ? attrs : {});
-export const a = (attrs: { href: string; title?: string }) =>
+export const a = (attrs: LinkAttributes) =>
   markFactory(sampleSchema.marks.link, attrs);
 export const emojiQuery = markFactory(sampleSchema.marks.emojiQuery, {});
 export const textColor = (attrs: { color: string }) =>