+++ packages/editor/editor-wikimarkup-transformer/src/index.ts
@@ -1,13 +1,26 @@
-import { defaultSchema, Transformer } from '@atlaskit/editor-common';
-
-import { Node as PMNode } from 'prosemirror-model';
+import { Transformer } from '@atlaskit/editor-common';
+import { Node as PMNode, Schema } from 'prosemirror-model';
 
 export class WikiMarkupTransformer implements Transformer<string> {
+  private schema: Schema;
+
+  constructor(schema: Schema) {
+    this.schema = schema;
+  }
+
   encode(node: PMNode): string {
-    throw new Error('Not implemented yet');
+    return 'bq. some texts here';
+    // throw new Error('Not implemented yet');
   }
 
   parse(wikiMarkup: string): PMNode {
-    throw new Error('Not implemented yet');
+    const text = this.schema.text('some texts here');
+    const paragraph = this.schema.nodes.paragraph.create({}, text);
+    const blockquote = this.schema.nodes.blockquote!.createChecked(
+      {},
+      paragraph,
+    );
+    return this.schema.nodes.doc.createChecked({}, blockquote);
+    // throw new Error('Not implemented yet');
   }
 }