+++ packages/editor/renderer/src/email/nodes/table-row.ts
@@ -1,0 +1,6 @@
+import { createTag } from '../util';
+import { NodeSerializerOpts } from '../interfaces';
+
+export default function tableRow({ text }: NodeSerializerOpts) {
+  return createTag('tr', {}, text);
+}