+++ packages/editor/renderer/src/react/nodes/mediaGroup.tsx
@@ -30,21 +30,23 @@ export default class MediaGroup extends PureComponent<
   render() {
     const numChildren = React.Children.count(this.props.children);
 
+    let content;
     if (numChildren === 1) {
       const card = React.Children.toArray(
         this.props.children,
       )[0] as ReactElement<any>;
       switch (card.props.type) {
         case 'file':
-          return this.renderSingleFile(card);
-
+          content = this.renderSingleFile(card);
+          break;
         case 'link':
         default:
-          return this.renderSingleLink(card);
+          content = this.renderSingleLink(card);
       }
     } else {
-      return this.renderStrip();
+      content = this.renderStrip();
     }
+    return <div className="MediaGroup">{content}</div>;
   }
 
   renderSingleFile(child: ReactElement<MediaProps>) {