+++ packages/media/media-picker/src/popup/components/views/upload/upload.tsx
@@ -7,18 +7,14 @@ import {
   CardView,
   CardEvent,
   OnLoadingChangeState,
-} from '@atlaskit/media-card';
-import {
-  Context,
-  CardActionType,
-  CardEventHandler,
-  MediaItem,
-  FileDetails,
   CardAction,
-} from '@atlaskit/media-core';
+  CardEventHandler,
+} from '@atlaskit/media-card';
+import { Context, MediaItem, FileDetails } from '@atlaskit/media-core';
 
 import Spinner from '@atlaskit/spinner';
 import Flag, { FlagGroup } from '@atlaskit/flag';
+import AnnotateIcon from '@atlaskit/icon/glyph/media-services/annotate';
 import EditorInfoIcon from '@atlaskit/icon/glyph/error';
 
 import { Browser } from '../../../..';
@@ -46,8 +42,8 @@ import { Wrapper, SpinnerWrapper } from './styled';
 const createEditCardAction = (handler: CardEventHandler): CardAction => {
   return {
     label: menuEdit,
-    type: CardActionType.custom,
     handler,
+    icon: <AnnotateIcon label={menuEdit} size="small" />,
   };
 };